Don't restrict pushitem "filename" to basenames
In the push item schema, the filename field was restricted to holding
only the basename of files with no leading path components.

This turns out to be a mistake in the design because the tools
expected to use this library don't work that way, and do have a few
cases of storing filenames with leading paths.  Operator metadata
bundles are one example, such as:
hco-bundle-registry-container-v2.0.0-42/operator_manifests.zip

So, let's loosen up the schema and allow any string to be provided
as a filename.

Fixes #7